--- 김자경/gdsc07/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'gdsc07'
