--- 김자경/gdsc07/src/main/java/com/example/gdsc07/Gdsc07Application.java
@@ -1,11 +1,0 @@
-package com.example.gdsc07;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Gdsc07Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Gdsc07Application.class, args);
-    }
-}
