--- 김자경/gdsc07/src/main/java/com/example/gdsc07/entity/Role.java
@@ -1,5 +1,0 @@
-package com.example.gdsc07.entity;
-
-public enum Role {
-    ROLE_ADMIN, ROLE_USER // ROLE_ 로 시작해야 인식
-}
