+++ 이현복/src/main/java/com/example/security2/Security2Application.java
@@ -1,0 +1,13 @@
+package com.example.security2;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Security2Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Security2Application.class, args);
+	}
+
+}
