+++ 이현복/src/main/java/com/example/security2/entity/Role.java
@@ -1,0 +1,9 @@
+package com.example.security2.entity;
+
+
+// Role : 시스템을 사용하는 사용자의 역할
+public enum Role {
+    
+    // ADMIN 권한, USER 권한 생성
+    ROLE_ADMIN, ROLE_USER
+}
