+++ 이현복/src/main/java/com/example/security2/repository/UserRepository.java
@@ -1,10 +1,13 @@
-package com.example.gdsc07.repository;
+package com.example.security2.repository;
 
-import com.example.gdsc07.entity.User;
+import com.example.security2.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-// username을 통해 사용자를 찾는 findByUsername 메소드
 public interface UserRepository extends JpaRepository<User, Long> {
+
+
+    // User이름을 찾을 수 있도록 하는 findByUsername
     Optional<User> findByUsername(String username);
 }
